refactor: unify PostCard component
