+++ src/components/feed/Index.ts
@@ -1,4 +1,4 @@
 // src/components/feed/index.ts
 export { default } from "./Feed";
 export { default as Feed } from "./Feed";
-export { default as PostCard } from "./PostCard";
+export { default as PostCard } from "../PostCard";
