--- src/components/feed/postcard.css
@@ -1,196 +1,0 @@
-/* ONE glass frame around the media; tiny radii; icon row inside the same frame.
-   No dark separator bar. Mobile-first, high contrast on dark scenes. */
-
-:root{
-  --pc-radius: 4px;                     /* tiny, consistent fillet */
-  --pc-glass: rgba(255,255,255,.08);    /* frosted white on dark background */
-  --pc-stroke: rgba(255,255,255,.14);   /* hairline */
-  --pc-ink: #eef2ff;                    /* icons/text on glass */
-  --pc-muted: rgba(255,255,255,.65);
-  --pc-ring: rgba(106,115,255,.22);     /* focus/hover inner ring */
-  --pc-circle-size: 2.375rem;
-  --pc-icon-size: 1.375rem;
-  --pc-btn-height: 2.25rem;
-  --pc-comment-close-size: 1.75rem;
-}
-
-/* article */
-.pc{
-  position: relative;
-  width: 100%;
-  margin: 12px auto;
-}
-
-/* ————— Glass Frame ————— */
-.pc-frame{
-  position: relative;
-  padding: 14px 10px 10px;              /* top a bit thicker for info */
-  background: var(--pc-glass);
-  border: 1px solid var(--pc-stroke);
-  border-radius: var(--pc-radius);
-  color: var(--pc-ink);
-  backdrop-filter: blur(16px) saturate(160%);
-  -webkit-backdrop-filter: blur(16px) saturate(160%);
-  box-shadow:
-    0 10px 28px rgba(0,0,0,.28),
-    inset 0 1px 0 rgba(255,255,255,.06);
-}
-
-/* header inside the glass */
-.pc-head{
-  display:flex; align-items:center; gap:10px;
-  margin-bottom: 10px;
-}
-.pc-ava{
-  width:var(--pc-circle-size); height:var(--pc-circle-size); border-radius:999px; overflow:hidden;
-  border:1px solid var(--pc-stroke); background:#111;
-  display:grid; place-items:center; cursor:pointer; max-width:100%; box-sizing:border-box;
-}
-.pc-ava.small{ width:1.75rem; height:1.75rem; }
-.pc-ava img{ width:100%; height:100%; object-fit:cover; display:block; }
-.pc-meta{ display:flex; flex-direction:column; min-width:0; line-height:1.1; }
-.pc-handle{ font-weight:800; white-space:nowrap; overflow:hidden; text-overflow:ellipsis; }
-.pc-sub{ font-size:12px; color:var(--pc-muted); }
-.pc-title{
-  margin-left:auto; padding:6px 10px; border-radius:8px;
-  border:1px solid var(--pc-stroke);
-  background: rgba(255,255,255,.06);
-  font-weight:800; white-space:nowrap; max-width:45%;
-  overflow:hidden; text-overflow:ellipsis;
-}
-
-/* media (black panel) */
-.pc-media{
-  background:#0b0d12;
-  border-radius: var(--pc-radius);
-  overflow: hidden;
-}
-.pc-media img,
-.pc-media video,
-.pc-media iframe,
-.pc-media model-viewer,
-.pc-media canvas{
-  display:block; width:100%; height:auto;
-  background:#0b0d12;
-  opacity:0; transition:opacity .25s ease;
-}
-
-.pc-media iframe{
-  aspect-ratio: 4 / 5;
-  max-height: 80vh;
-}
-
-/* carousel (for multiple images) */
-.pc-carousel{
-  display:grid; grid-auto-flow:column; grid-auto-columns:100%;
-  overflow-x:auto; scroll-snap-type: x mandatory; gap:0;
-  -webkit-overflow-scrolling: touch;
-}
-.pc-carousel::-webkit-scrollbar{ display:none; }
-.pc-slide{ scroll-snap-align:start; }
-
-/* footer inside the glass */
-.pc-foot{
-  margin-top:10px;
-  display:flex; align-items:center; justify-content:space-between;
-}
-.pc-actions{ display:flex; align-items:center; gap:10px; flex-wrap:wrap; max-width:100%; box-sizing:border-box; }
-.pc-btn{
-  width:var(--pc-circle-size); height:var(--pc-circle-size); border-radius:999px;
-  border:1px solid var(--pc-stroke);
-  background: rgba(255,255,255,.06);
-  color: var(--pc-ink); display:grid; place-items:center;
-  transition: transform .12s ease, box-shadow .18s ease, border-color .15s ease; max-width:100%; box-sizing:border-box;
-}
-.pc-btn:hover{
-  transform: translateY(-1px);
-  border-color: rgba(106,115,255,.55);
-  box-shadow: 0 0 0 2px var(--pc-ring) inset;
-}
-.pc-ico{ width:var(--pc-icon-size); height:var(--pc-icon-size); display:block; max-width:100%; box-sizing:border-box; }
-
-/* “pop-out” single-row emoji strip, extending from the glass */
-.pc-react-strip{
-  display:flex; align-items:center; gap:8px; overflow-x:auto;
-  padding:6px 8px; border-radius:999px;
-  background: var(--pc-glass); border:1px solid var(--pc-stroke);
-  color: var(--pc-ink);
-  backdrop-filter: blur(16px) saturate(160%);
-  -webkit-backdrop-filter: blur(16px) saturate(160%);
-  margin-left:8px;
-}
-.pc-react-strip::-webkit-scrollbar{ display:none; }
-.pc-emo{
-  font-size:20px; line-height:1;
-  background:transparent; border:0; cursor:pointer;
-  padding:6px 8px;
-}
-
-.pc-comment-box{
-  margin-top:8px;
-  display:flex; flex-direction:column; gap:8px;
-  padding:8px 10px;
-  border-radius: var(--pc-radius);
-  background: var(--pc-glass); border:1px solid var(--pc-stroke);
-  color: var(--pc-ink);
-  backdrop-filter: blur(16px) saturate(160%);
-  -webkit-backdrop-filter: blur(16px) saturate(160%);
-}
-.pc-comments{ list-style:none; margin:0; padding:0; display:flex; flex-direction:column; gap:6px; max-height:100px; overflow:auto; }
-.pc-comments li{ background: rgba(255,255,255,.06); border:1px solid var(--pc-stroke); padding:4px 8px; border-radius: var(--pc-radius); }
-.pc-addcmt{ display:flex; gap:8px; }
-.pc-addcmt input{ flex:1; padding:6px 8px; border-radius: var(--pc-radius); border:1px solid var(--pc-stroke); background: rgba(0,0,0,.2); color: var(--pc-ink); max-width:100%; box-sizing:border-box; }
-.pc-addcmt button{ padding:6px 12px; border-radius: var(--pc-radius); border:1px solid var(--pc-stroke); background: rgba(255,255,255,.06); color: var(--pc-ink); height:var(--pc-btn-height); max-width:100%; box-sizing:border-box; }
-.pc-comment-close{ align-self:flex-end; width:var(--pc-comment-close-size); height:var(--pc-comment-close-size); border-radius:999px; border:1px solid var(--pc-stroke); background: rgba(255,255,255,.06); color: var(--pc-ink); max-width:100%; box-sizing:border-box; }
-.pc-empty{ opacity:.7; }
-
-.pc-remix-box{
-  margin-top:8px;
-  display:flex; flex-direction:column; gap:8px;
-  padding:8px 10px;
-  border-radius: var(--pc-radius);
-  background: var(--pc-glass); border:1px solid var(--pc-stroke);
-  color: var(--pc-ink);
-  backdrop-filter: blur(16px) saturate(160%);
-  -webkit-backdrop-filter: blur(16px) saturate(160%);
-}
-
-.pc-count{
-  margin-left:4px;
-  font-size:12px;
-  background: rgba(255,255,255,.06);
-  border:1px solid var(--pc-stroke);
-  padding:2px 6px;
-  border-radius:999px;
-}
-
-.pc-reactions{
-  margin-top:8px;
-  display:flex;
-  gap:6px;
-  flex-wrap:wrap;
-  color: var(--pc-ink);
-}
-.pc-emo-count{
-  font-size:14px;
-  background: rgba(255,255,255,.06);
-  border:1px solid var(--pc-stroke);
-  padding:2px 6px;
-  border-radius:999px;
-}
-
-/* focus state for accessibility */
-.pc-btn:focus-visible,
-.pc-ava:focus-visible,
-.pc-emo:focus-visible{
-  outline: none;
-  box-shadow: 0 0 0 2px var(--pc-ring) inset;
-}
-
-/* drag-target class (kept for the orb) */
-.pc-target .pc-frame{ outline: 3px solid rgba(255,116,222,.5); outline-offset:-3px; }
-
-/* small devices */
-@media (max-width: 420px){
-  .pc-title{ display:none; }             /* keep header compact on phones */
-}
